modernizing MacOS project for Xcode 5
* changed support from 10.6 to 10.7 - it has to happen at some point;
  why not now?

* removed support for 32 bit arch - there hasn't been one of these in
  a long time

* converted several NSInteger types to short to suppress compiler
  warnings iOS/MacOS

* suppressed compiler warnings in `GHTesting.m` with an internal
  category

* removed duplicate method definition
  receiveData:statusCode:MIMEType:afterDelay from GHMockURLConnection

* enabled many, many, many warnings to get a handle on the state of
  the code

+++ Classes/GHTest/GHTest.h
@@ -68,16 +68,16 @@ extern BOOL GHTestStatusEnded(GHTestStatus status);
  Test stats.
  */
 typedef struct {
-  NSInteger succeedCount; // Number of succeeded tests
-  NSInteger failureCount; // Number of failed tests
-  NSInteger cancelCount; // Number of aborted tests
-  NSInteger testCount; // Total number of tests 
+  short succeedCount; // Number of succeeded tests
+  short failureCount; // Number of failed tests
+  short cancelCount; // Number of aborted tests
+  short testCount; // Total number of tests
 } GHTestStats;
 
 /*!
  Create GHTestStats.
  */
-extern GHTestStats GHTestStatsMake(NSInteger succeedCount, NSInteger failureCount, NSInteger cancelCount, NSInteger testCount);
+extern GHTestStats GHTestStatsMake(short succeedCount, short failureCount, short cancelCount, short testCount);
 
 extern const GHTestStats GHTestStatsEmpty;
 
@@ -164,7 +164,7 @@ extern NSString *NSStringFromGHTestStats(GHTestStats stats);
 /*!
  @result The number of disabled tests
  */
-- (NSInteger)disabledCount;
+- (short)disabledCount;
 
 @end
